Show arrows on solid sankey edges
The circular sankey layout isn't great at sizing to fit the existing space and this corrects it by adjusting the sankey edges and nodes in the zoom phase in network frame. But it adjusts the datapoints themselves and d3-circular-sankey has already calculated the edge `d` values for SVG `path` elements.

The edges drawn by `d3-sankey-circular` are meant to be paths where stroke-width encodes the flow, which has its strengths and weaknesses. The edges in Semiotic are areas. The main problem is that you can't get arrows drawn over areas (because the arrow code is based on the line which for stroke-encoding is the middle of the path but for area encoding it's the perimeter) so this adds an option to `networkSettings` called `showArrows` and draws arrows on top of the paths if it's set to true.